+++ pkg/kubeapiserver/authenticator/BUILD
@@ -31,6 +31,7 @@ go_library(
         "//staging/src/k8s.io/apiserver/plugin/pkg/authenticator/token/webhook:go_default_library",
         "//staging/src/k8s.io/client-go/plugin/pkg/client/auth:go_default_library",
         "//staging/src/k8s.io/client-go/util/cert:go_default_library",
+        "//staging/src/k8s.io/client-go/util/keyutil:go_default_library",
         "//vendor/github.com/go-openapi/spec:go_default_library",
     ],
 )
