+++ staging/src/k8s.io/client-go/util/keyutil/OWNERS
@@ -1,0 +1,7 @@
+approvers:
+- sig-auth-certificates-approvers
+reviewers:
+- sig-auth-certificates-reviewers
+labels:
+- sig/auth
+
